Move changedFiles logic into FindFiles
There wasn't really much happening in FindFiles, so the unexported
function code was moved to eliminate a level of indirection. This also makes
it cleaner to access the options for the command.
