+++ src/gitsh.go
@@ -43,45 +43,6 @@ func commitTimeStamp(obj string) (string, error) {
 	return strings.Trim(line, "\""), nil
 }
 
-// changedFiles returns the paths of all files changed in commits between
-// master and the current branch.
-// iExt is a list of file extensions to use to filter final results.
-// iPaths is a list of path prefixes to use to filter final results.
-func changedFiles(iExt []string, iPaths []string) ([]string, error) {
-	var lines []string
-	out, err := run("git diff master HEAD --name-only")
-
-	if err != nil {
-		return lines, err
-	}
-
-	for _, line := range strings.Split(out, "\n") {
-		l := strings.Trim(line, " ")
-
-		passExtCheck := true
-		if len(iExt) > 0 {
-			for _, ext := range iExt {
-				passExtCheck = passExtCheck && !strings.HasSuffix(line, ext)
-			}
-		}
-
-		passPathCheck := true
-		lLen := len(line)
-		if len(iPaths) > 0 {
-			for _, prefix := range iPaths {
-				passPathCheck = passPathCheck &&
-					len(strings.TrimPrefix(line, prefix)) == lLen
-			}
-		}
-
-		if len(l) > 0 && passExtCheck && passPathCheck {
-			lines = append(lines, l)
-		}
-	}
-
-	return lines, err
-}
-
 // committerCounts finds recent committers and commit counts for
 // the file at `path`. It uses 2 channels to communicate the state of
 // processing. If `since` is a proper 'YYYY-MM-DD' formatted date, the
